Moved sequence class and generator to a dedicated package
