+++ src/main/java/pl/edu/pw/sportyapp/shared/sequence/SequenceGeneratorService.java
@@ -1,4 +1,4 @@
-package pl.edu.pw.sportyapp.shared;
+package pl.edu.pw.sportyapp.shared.sequence;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoOperations;
